requirements.txt
paramiko==3.4.0
python-dotenv